Refactor info-files command: separate CLI from display logic
✨ NEW STRUCTURE:
- cmd/treex/cmd/info_files.go: Clean CLI interface only
- Self-contained help display logic in dedicated file

🔧 IMPROVEMENTS:
- Better command organization with dedicated file
- Cleaner separation from root.go monolith
- Enhanced CLI pattern consistency
- Self-contained help functionality

📦 CHANGES:
- Moved info-files command definition to separate file
- Removed help display logic from root.go
- Used RunE for consistent error handling pattern
- Maintained all existing functionality

✅ TESTING:
- All tests pass (51 tests)
- Command functionality verified
- Help display works correctly
- No regressions introduced

Simple but effective separation improving code organization.
